# Heavily improve customization performance